added Integration test with Liquibase

+++ src/main/resources/db/changelog/changeset/create-table-users.sql
@@ -2,5 +2,5 @@ create table users (
     id BIGSERIAL PRIMARY KEY ,
     username VARCHAR(64) ,
     name VARCHAR (64) ,
-    company_id BIGINT REFERENCES users(id)
+    company_id INT REFERENCES users(id)
 );
